% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmCD.R
\docType{data}
\name{dmCD}
\alias{dmCD}
\title{Pre-processed dendrometer data from Camp Daniel, Canada}
\format{A \code{data.frame} with a timestamp (\code{\%Y-\%m-\%d \%H:\%M:\%S} format) as row names, and the dendrometer series in the first column.}
\usage{
data(dmCD)
}
\description{
This dataset presents a pre-processed version of \code{\link{dmCDraw}}, in which different time variables were converted to a timestamp using code provided in the \code{Import dendrometer data} vignette.
}
\keyword{datasets}

