% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmCDraw.R
\docType{data}
\name{dmCDraw}
\alias{dmCDraw}
\title{Raw dendrometer data from Camp Daniel, Canada}
\format{A \code{data.frame} with a dendrometer series and various time variables.}
\source{
Sergio Rossi
}
\usage{
data(dmCDraw)
}
\description{
This dataset presents hourly dendrometer series for a black spruce (\emph{Picea mariana} (Mill.) BSP) tree from Camp Daniel, Canada, for the year 2008.
}
\keyword{datasets}

