% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmEDraw.R
\docType{data}
\name{dmEDraw}
\alias{dmEDraw}
\title{Raw dendrometer data from Eldena, Germany}
\format{A \code{data.frame} with dendrometer series and a timestamp.}
\source{
Martin Wilmking
}
\usage{
data(dmEDraw)
}
\description{
This dataset presents half-hourly dendrometer series for two European beech (\emph{Fagus sylvatica} L.) trees from the monitoring plot Eldena, Germany, for the year 2015.
}
\keyword{datasets}

