% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmHS.R
\docType{data}
\name{dmHS}
\alias{dmHS}
\title{Pre-processed dendrometer data from Hinnensee, Germany}
\format{A \code{data.frame} with a timestamp (\code{\%Y-\%m-\%d \%H:\%M:\%S} format) as row names, and the dendrometer series in the first column.}
\usage{
data(dmHS)
}
\description{
This dataset presents a pre-processed version of \code{\link{dmHSraw}}, using code provided in the \code{Import dendrometer data} vignette.
}
\keyword{datasets}

