% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmHSraw.R
\docType{data}
\name{dmHSraw}
\alias{dmHSraw}
\title{Raw dendrometer data from Hinnensee, Germany}
\format{A \code{data.frame} with a dendrometer series and various time variables.}
\source{
Sonia Simard
}
\usage{
data(dmHSraw)
}
\description{
This dataset presents half-hourly dendrometer series for a European beech (\emph{Fagus sylvatica} L.) tree from the monitoring plot Hinnensee, Germany, for the year 2012.
}
\keyword{datasets}

