% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envED.R
\docType{data}
\name{envED}
\alias{envED}
\title{Environmental data from Eldena, Germany}
\format{A \code{data.frame} with a timestamp (\code{\%Y-\%m-\%d \%H:\%M:\%S} format) as row names, and air and soil temperature parameters in columns.}
\source{
Martin Wilmking
}
\usage{
data(envED)
}
\description{
This dataset presents some temperature data from the monitoring plot Eldena, Germany, for the year 2015.
}
\keyword{datasets}

