% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.dendro.R
\name{is.dendro}
\alias{is.dendro}
\title{Check input data}
\usage{
is.dendro(dm.data)
}
\arguments{
\item{dm.data}{a \code{data.frame} with a timestamp (\code{\%Y-\%m-\%d \%H:\%M:\%S} format) as row names, and dendrometer series in columns. Output as created using code from the \code{Import dendrometer data} vignette.}
}
\value{
The function returns \code{TRUE} if the input data is valid and \code{FALSE} otherwise. In the latter case, specific error messages are given as well.
}
\description{
The function checks whether the input data is in the required format, as described in the \code{Import} \code{dendrometer} \code{data} vignette.
}
\examples{
data(dmCD, dmHS, dmED)
is.dendro(dmCD)
is.dendro(dmHS)
is.dendro(dmED)

}
\author{
Ernst van der Maaten, Marieke van der Maaten-Theunissen and Marko Smiljanic.
}

